move uninstall logic to new file
